// apps/admin/src/lib/api/generated/models/topicImport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.6
 */
import type { TopicImportFormat } from './topicImportFormat';

export interface TopicImport {
  /**
     * @minLength 1
     * @maxLength 262144
     */
  content: string;
  format: TopicImportFormat;
  /**
     * @minLength 1
     * @maxLength 200
     */
  source_name: string;
}